usando usePokemondetail para el modal
